# Allow precomputing spike trains