--- a/src/spikeinterface/core/basesorting.py
+++ b/src/spikeinterface/core/basesorting.py
@@ -9,6 +9,13 @@ from .base import BaseExtractor, BaseSegment
 from .core_tools import spike_vector_to_dict
 from .waveform_tools import has_exceeding_spikes
 
+try:
+    import numba
+
+    HAVE_NUMBA = True
+except:
+    HAVE_NUMBA = False
+
 
 minimum_spike_dtype = [("sample_index", "int64"), ("unit_index", "int64"), ("segment_index", "int64")]
 
@@ -440,6 +447,9 @@ class BaseSorting(BaseExtractor):
         unit_ids = self.unit_ids
 
         if from_spike_vector:
+            assert HAVE_NUMBA, "`numba` must be installed to use `precompute_spike_trains(from_spike_vector=True)`\
+            Either install numba (pip install numba) or set `from_spike_vector=False`"
+
             spike_trains = spike_vector_to_dict(self.to_spike_vector())
 
             for segment_index in range(self.get_num_segments()):