--- a/src/spikeinterface/core/core_tools.py
+++ b/src/spikeinterface/core/core_tools.py
@@ -22,6 +22,13 @@ from .job_tools import (
     _shared_job_kwargs_doc,
 )
 
+try:
+    import numba
+
+    HAVE_NUMBA = True
+else:
+    HAVE_NUMBA = False
+
 
 def define_function_from_class(source_class, name):
     "Wrapper to change the name of a class"
@@ -938,20 +945,23 @@ def convert_bytes_to_str(byte_value: int) -> str:
 
 
 def spike_vector_to_dict(spike_vector: np.ndarray) -> dict:
+    assert HAVE_NUMBA
+
     spike_trains = _vector_to_dict(spike_vector["sample_index"].astype(np.int64), spike_vector["unit_index"].astype(np.int64), spike_vector["segment_index"].astype(np.int64))
     
     return [{unit_index: np.array(spike_trains[seg][unit_index]) for unit_index in spike_trains[seg].keys()} for seg in range(len(spike_trains))]
 
-@numba.jit((numba.int64[::1], numba.int64[::1], numba.int64[::1]), nopython=True, nogil=True, cache=True)
-def _vector_to_dict(sample_index, unit_index, segment_index):
-    spike_trains = numba.typed.List()
+if HAVE_NUMBA:
+    @numba.jit((numba.int64[::1], numba.int64[::1], numba.int64[::1]), nopython=True, nogil=True, cache=True)
+    def _vector_to_dict(sample_index, unit_index, segment_index):
+        spike_trains = numba.typed.List()
 
-    for seg in range(1 + np.max(segment_index)):
-        spike_trains.append(numba.typed.Dict())
-        for i in range(1 + np.max(unit_index)):
-            spike_trains[seg][i] = numba.typed.List.empty_list(numba.int64)
+        for seg in range(1 + np.max(segment_index)):
+            spike_trains.append(numba.typed.Dict())
+            for i in range(1 + np.max(unit_index)):
+                spike_trains[seg][i] = numba.typed.List.empty_list(numba.int64)
 
-        for i in range(len(sample_index)):
-            spike_trains[seg][unit_index[i]].append(sample_index[i])
+            for i in range(len(sample_index)):
+                spike_trains[seg][unit_index[i]].append(sample_index[i])
 
-    return spike_trains
+        return spike_trains