… src/spikeinterface/core/tests/test_core_tools.py
… src/spikeinterface/core/tests/test_sorting_tools.py