Allows for fast pre-computation of spike trains