@DradeAW 2023-11-06
Wow, to my surprise this is not faster!
I don't understand why, maybe I made a mistake somewhere ...

```python
import time
import spikeinterface.core as si


t1 = time.perf_counter()

sorting = si.load_extractor("/mnt/raid0/data/MEArec/1h_3000cells/analyses/ks2_5_pj7-3/sorting")
for unit_id in sorting.unit_ids:
	sorting.get_unit_spike_train(unit_id)

t2 = time.perf_counter()

sorting = si.load_extractor("/mnt/raid0/data/MEArec/1h_3000cells/analyses/ks2_5_pj7-3/sorting")
sorting.precompute_spike_trains()
for unit_id in sorting.unit_ids:
	sorting.get_unit_spike_train(unit_id)

t3 = time.perf_counter()

print(f"Old way: {t2-t1:.1f} s")
print(f"New way: {t3-t2:.1f} s")
```

```
Old way: 16.5 s
New way: 18.3 s
```