@DradeAW 2023-11-06
Ok this is way better with the latest commit!

```python
import time
import numpy as np
import spikeinterface.core as si



sorting1 = si.load_extractor("/mnt/raid0/data/MEArec/1h_3000cells/analyses/ks2_5_pj7-3/sorting")
sorting2 = si.load_extractor("/mnt/raid0/data/MEArec/1h_3000cells/analyses/ks2_5_pj7-3/sorting")

t1 = time.perf_counter()

for unit_id in sorting1.unit_ids:
	sorting1.get_unit_spike_train(unit_id)

t2 = time.perf_counter()

sorting2.precompute_spike_trains()
for unit_id in sorting2.unit_ids:
	sorting2.get_unit_spike_train(unit_id)

t3 = time.perf_counter()

print(f"Old way: {t2-t1:.1f} s")
print(f"New way: {t3-t2:.1f} s")

for unit_id in sorting1.unit_ids:
	assert np.all(sorting1.get_unit_spike_train(unit_id) == sorting2.get_unit_spike_train(unit_id))
```

```
Old way: 16.3 s
New way: 1.4 s
```