@DradeAW 2023-11-06
For reference, my sorting dataset contains 531 units for a total of 11,298,565 spikes (1h recording)