@samuelgarcia 2023-11-06
Hi Aurelien.
thanks for this.

I will not have time to get feedback on this very soon.
But globaly:
 * thanks a lot for this speed up, this was on my todo list
 * not sure to like the idea of having numba in core... lets see. It can be used optionally if installed, but then we need
    it for  testing.
* not sure to like the semantic in one function. In short many Sorting are already spiketrains centric (most of then) and 
   some of then are spike train first (numpysorting, ...). We need this function for 2 clear distinct case:
   1. when spike train centric to preload in memory
   2. when spike vector centric to make the convertion all at once. The caching used to be spiketrain per spiketrain which 
      also a valid scenario I think.
  The semantic have to be very clear for this 2 cases. This is now done with one function with one option.
* I like the idea to explicitly call this function.