@DradeAW 2023-11-07
> not sure to like the semantic in one function [...] We need this function for 2 clear distinct case

@samuelgarcia You would prefer to split this in two functions with no parameter each?