@DradeAW 2023-11-07
> not sure to like the idea of having numba in core... lets see. It can be used optionally if installed, but then we need it for testing.

I removed the need for numba (there is now a default if it's not installed).
For the tests, I've added a skip if numba is not installed, so it will not make the test fail.