@DradeAW 2023-11-09
> How does this interacts plays with the caching option in `to_spike_vector` and `get_unit_spike_train`.

It overrides it (forces a recompute).

> Isn't it duplicating behavior?

No, because it is a fast computation of all spike trains at once, rather than one unit by one unit (which is, as I showed above, much faster).

> How would this play out with multiprocessing, all this pre-computing is lost isn't it?

I'm not sure, but I believe if you get the spike train it should find it in the cache in all cases?