@h-mayorquin 2023-11-09
> No, because it is a fast computation of all spike trains at once, rather than one unit by one unit (which is, as I showed above, much faster).

So are the functions introduced here doing a better job at calculating the `spike_vector` than `to_spike_vector`?