@DradeAW 2023-11-09
> So are the functions introduced here doing a better job at calculating the spike_vector than to_spike_vector?

No, I haven't touched that.
This is in converting the `spike_vector` into spike trains.
The method right now loops over all units, and calls `spike_train = spike_vector[ids == unit_id]`
Which is very slow for big vectors.

This new method allows the conversion in a very efficient manner!