@h-mayorquin 2023-11-09
> No, I haven't touched that.
This is in converting the spike_vector into spike trains.
The method right now loops over all units, and calls spike_train = spike_vector[ids == unit_id]
Which is very slow for big vectors.

Got you,thanks, I realized I was not understand what you were doing and went and re-read again.