@DradeAW 2023-11-15
Here is the benchmark with a bigger sorting object (the biggest one I ever used):
3,000 units, 1 hour, total of 74,638,923 spikes (same code as shown above, but with a different sorting object)

```
Old way: 614.1 s
New way: 7.7 s
```

This is ~80 times faster.