@h-mayorquin 2023-11-15
Yeah, I don't really like having numba for something like this in the core, can you profile against this vectorized version for the last example above?

```python
    def precompute_spike_trains(self):
        """
        Pre-computes and caches all spike trains for this sorting. Should be 
        quicker if you already have the spike vector.

        """
    
        unit_ids = self.unit_ids

        spike_trains = {}
        
        spike_vector_per_segment = self.to_spike_vector(concatenated=False)
        for segment_index, spike_vector in enumerate(spike_vector_per_segment):
            spike_trains_segment = dict()
            for unit_id in unit_ids:
                spike_train = spike_vector[spike_vector['unit_index'] == unit_id]['sample_index']
                spike_trains_segment[unit_id] = spike_train

            spike_trains_segment[segment_index] = spike_trains_segment

        self._cached_spike_trains = spike_trains
```