@samuelgarcia 2023-11-17
@zm711 @DradeAW @h-mayorquin 
I made a refactor of this that should made happy all of us (included me)

* `create a sorting_tools.py`
* cache from spike vector or not is automatic depending spike_vector is cached or not
* the function import numba internally in a try cache. If numba is not there then have another implementation
* use the none concatenated version of spike vector which is faster.
* have explicit num_units rather than max(unit_index) + 1
* return directly a dict of dict

Tell me what you think.