@DradeAW 2023-11-17
There are probably a few places in `core` that could be improved by calling this function, because they need the spike train of all unit ids (maybe when extracting waveforms?)