@DradeAW 2023-11-17
Latest benchmark: even faster

~200x times faster on a very big sorting:
```
Old way: 634.3 s
New way: 2.9 s
```

~35 times faster for a sorting I typically use:
```
Old way: 24.9 s
New way: 0.7 s
```