@h-mayorquin 2023-11-18
Thanks @samuelgarcia . I introduced some machinery to test this:

https://github.com/SpikeInterface/spikeinterface/pull/2227

I will give it a  test as soon as I can.